feat: refactor data imports and replace with test data files; remove old data files
