+++ package.json
@@ -10,6 +10,7 @@
     "@fastify/swagger": "^9.4.2",
     "@fastify/swagger-ui": "^5.2.2",
     "fastify": "^5.2.2",
+    "pg": "^8.14.1",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
